move xstate shared
