+++ workflows/sample/add-tests-xstate.test.ts
@@ -1,7 +1,7 @@
 import { AddTestsWorkflowMachine } from "./add-tests-xstate.ts";
 import { createActor, waitFor } from "xstate";
 import { describe, it, expect } from "vitest";
-import { allChildrenSettled } from "./xstate-shared.ts";
+import { allChildrenSettled } from "../src/xstate-shared.ts";
 
 describe("AddTestsWorkflow", () => {
   it("should print instructions", async () => {
@@ -9,8 +9,6 @@ describe("AddTestsWorkflow", () => {
       input: { path: "sample/add-tests-xstate.ts" },
     });
     actor.start();
-    console.log("all good?", actor.getSnapshot().status);
-    console.log(actor.getSnapshot());
     await waitFor(actor, allChildrenSettled);
     expect(actor.getSnapshot().value).toBe("addingTests");
     actor.send({ type: "continue" });
