+++ workflows/src/workflow.ts
@@ -8,7 +8,7 @@ import type {
 import type { AnyStateMachine, AnyActor } from "xstate";
 import { addNewLinesToString } from "./utils.ts";
 import { createActor, waitFor } from "xstate";
-import { allChildrenSettled } from "../sample/xstate-shared.ts";
+import { allChildrenSettled } from "./xstate-shared.ts";
 // The following is TS magic to describe a class constructor that implements the abstract SimpleWorkflow class.
 type AbstractClassConstructor<T extends Workflow> = new (...args: any[]) => T;
